// src/app/models/arac.model.ts
export interface AracModel {
  id: any;
  category: any;
  categoryId: any;
  state: any;
  stateId: any;
  name: string;
  model: string;
  year: number;
  price: number;
  km: number;
  engine_power: string;
  engine_volume: string;
  color: string;
  imageUrl: string;
  imageData: string;
  description: string;
  isActive: boolean;
  userId: any;
  userName: any;
}